Enforce that each event must record the actor that caused it
Tried to hide as much of this from Command['process'] as possible by
having a 'system' fallback actor and tranparently passing it through
from the POST handlers. The one place it does surface is in tests of the
'process' as the commands need to be augmented with an actor. There is
now an 'arbitraryActor()' helper for tests to help with this.