Add item selection component
Adds a new handlebar template which maps over an array of elements
and turns them into a gridjs filterable table. Use this new component
to improve the /equipment/mark-member-trained page.

This component can be used as follows:

{{#filterList members "Members"}}
  {{this.name}}
{{/filterList}}

This is roughly equivalent to:

<table>
  <thead><tr><th>Members</th></tr></thead>
  <tbody>
    <tr><td>Alice</td></tr>
    <tr><td>Bob</td></tr>
  </tbody>
</table>

+ gridjs integration.

Also reduce the amount of boilerplate required for gridjs. Previously
each use required a manually specified wrapper div where the gridjs
table would be inserted as well as a script tag instantiating the table:

<div id="wrapper"></div>
<table id="members">
  <thead>...</thead>
  <tbody>...</tbody>
</table>
<script>
  new GridJs(...).render(...);
</script>

This has been replaced with a single common script tag embeded in the
page template which looks for all tagged tables and creates a gridjs
grid for it. Now the code looks like:

<table data-gridjs>
  <thead>...</thead>
  <tbody>...</tbody>
</table>