docker-compose -> docker compose
Switch from using docker-compose to docker compose.
With the latter I was getting odd behavior when trying to stop the
container with Ctrl-C:

$ docker-compose --file docker-compose.yaml --file docker-compose.dev.yaml up --build
[...]
members-app-app-1          | INFO [19:24:22.795] (30): Visit http://localhost:8080 to see the application
members-app-app-1          | INFO [19:24:22.795] (30): Visit http://localhost:1080 to see the emails it sends
^CGracefully stopping... (press Ctrl+C again to force)
[+] Running 3/3
 ⠿ Container members-app-sync_worker-1  Stopped                                                                                                                                 0.6s
 ⠿ Container members-app-app-1          Stopped                                                                                                                                 0.6s
 ⠿ Container members-app-mailcatcher-1  Stopped                                                                                                                                 0.1s
^Cno container to kill^C^C^C^C^C^C

I don't get the same with `docker compose` which is theoretically a)
mostly the same and b) the new better replacement for docker-compose.