add adhoc shared-read-model queries for superusers
Frequently there are questions like:

"Who has done the most trainings?"
"Which equipment has the most trained members?"
"How many owners are there?"

Which would be easy to answer using the read model using queries like:

select trainedByMemberNumber, count(*)
from trainedMembers
group by trainedByMemberNumber
order by 2 desc;

This adds an adhoc query page for the shared-read-model at /db to allow
this. The query page is:
- gated to superusers
- read-only
and operates on the read model.

In addition this prototypes an approach to writing handlers in a
more traditional imperative style. The hope is this would be more
approachable for new contributors. This is done using an adaptor which
takes an imperative handler and adapts it into the normal fp-ts 'Query'
type allowing the experiment to be self contained to this single
endpoint.