Add datetime selecter for trained time for trainers (not just admins)
The datetime picker is restricted to within one month.
Note: This does re-use the MarkMemberTrainedBy event and sets the marked
trained by to the trainer (only allowed if less than one month since
training). This is a slight deviation from the current meeaning of
MarkMemberTrainedBy as that currently would always be an dmin.key.